Add tileset image dimension checks
This checks for valid input bitmaps sizes, which must be a single column
of 32x32 pixel tiles. This implies a width of 32 pixels, and a height
which is a multiple of 32 pixels.

This also checks that after scaling, the scaled image will have
dimensions that fit within the memory size.